Rewrite Inventory
Rewrite the inventory type to only allow items to be added to certain slots.
An inventory will have the null item in each slot when it is created. When an
item is added to a slot, the null item in that slot is replaced with the
added item. This will enforce the idea that inventories have a certain
capacity that cannot be changed.