Rewrite Entities Package
The Sakura Entities package will operate on the concept of prefabs
representing "classes" and instances of prefabs representing "instances" of
those "classes". The scripts on a prefab "class" can be either "constructors",
which are scripts that initialize the "class" instance, or "methods", which
are scripts that represent a single behavior of that "class". These "methods"
can be called by creating Unity events that enable one of these "method"
script components on a given "instance".