Modify Shipping Bin UI
Modify the shipping bin UI to use two instances of the InventoryPanel prefab,
one for each of the sell and keep sections. This will allow for the sell and
keep sections to be set to displaying different inventories. The sell section
will display the shipping bin's inventory while the keep section will display
the player's inventory.